Add edit button next to each task
Double clicking a label is not as intuitive to edit a task,
so adding a button would help a user know that they can
edit.

Had to find where the delete button was and insert the edit button
next to it. Then look in todo-view to add in the functionality
when the mouse clicks on the edit button.