Add priority feature
Added a button in the template file to indicate to the user that they
can click on it if they want priority on a certain task. In the todo-
view, had to link the button with a toggle function. Also
in the models file, it saves the data whether the label's
has priority or not.